Add files via upload
Use the typical finance definition of hedge:

To hedge, in finance, is to take an offsetting position in an asset or investment that reduces the price risk of an existing position. A hedge is therefore a trade that is made with the purpose of reducing the risk of adverse price movements in another asset. Normally, a hedge consists of taking the opposite position in a related security or in a derivative security based on the asset to be hedged.

Here are a few suggestions:

Measure Bitcoin's price with BTC-USD and gold's price with GLD
Throughout the project, assume Bitcoin and U.S. public equity markets have the same closing time
Measure the price level with PCEPI from the Federal Reserve Database (FRED), which is downloadable with pdr.DataReader()
Measure inflation (i.e., the rate of change in the price level) as the percent change in PCEPI